requirements.txt
flask>=2.3.0,<3.0.0
requests>=2.31.0,<3.0.0
python-dotenv>=1.0.0,<2.0.0
matplotlib>=3.7.0,<4.0.0
